Create make_eco_phylo_rasters.R
The creation of the rasters isn't currently working. The first layer of eBird abundance data that's loaded in (for the Greater Rhea) appears to be a raster of NaN's. We need to explore the eBird abundance data more carefully to work out if this is a problem with this specific species or with how we load in the data.